appsetting  json file modified

+++ TechBoxtoUnbox/Models/DTO/Course.cs
@@ -5,11 +5,11 @@ namespace TechBoxtoUnbox.Models.DTO
     public class Course
     {
         [Key]
-        public int CourseId { get; set; }
+        public int CourseId { get; set; } 
         
         public string CName { get; set; }
         
-        public decimal CFee { get; set; }
+        public double CFee { get; set; }
         public DateTime Createddate { get; set; }
     }
 }
